Updated game objects for new camera system.
All GameObject children now implement IsOnScreen as a property instead of a method. The reference to an offset vector outside of Update() (as well as any need for a translation matrix) is eliminated by having each object keep track of their screen position, which is updated in Update() using the passed in offset vector. Note: Game1 still has errors, so the program will not currently run!