Enemies properly follow the player now.
The direction vector was previously being normalized even if it was the zero vector, which caused the teleportation. There's a check now that fixes the problem. Note: enemies still won't always show up properly (this seems to be an IsOnScreen issue). Also, god/debug mode can now be toggled on and off with the G key.